+++ src/components/ClusterNodes/ClusterNodeContainer.tsx
@@ -1,14 +1,11 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { URLS } from '../../config'
 import { Redirect, Route, Switch } from 'react-router-dom'
 import ClusterList from './ClusterList'
 import NodeDetails from './NodeDetails'
 import NodeList from './NodeList'
-import { MultiValue } from 'react-select'
-import { columnMetadataType } from './types'
 
 export default function ClusterNodeContainer() {
-    const [appliedColumns, setAppliedColumns] = useState<MultiValue<columnMetadataType>>([])
     return (
         <>
             <Switch>
@@ -16,7 +13,7 @@ export default function ClusterNodeContainer() {
                     <ClusterList />
                 </Route>
                 <Route path={`${URLS.CLUSTER_LIST}/:clusterId`} exact>
-                    <NodeList appliedColumns={appliedColumns} setAppliedColumns={setAppliedColumns} />
+                    <NodeList />
                 </Route>
                 <Route path={`${URLS.CLUSTER_LIST}/:clusterId/:nodeName`} exact>
                     <NodeDetails />
